+++ docs/about/history.rst
@@ -25,6 +25,8 @@ Changed
 .. bullet_list::
 
    -  Added new CPU-based collision provider - *Baked FFT Data*.
+   -  Add support for rendering in edit mode (camera preview and scene view) to *Underwater Renderer*.
+      It can be enabled/disabled with the fog scene view toggle.
    -  Add *CREST_OCEAN* scripting defines symbol.
    -  Add *Depth Fog Density Factor* to *Underwater Renderer* which can be used to decrease underwater fog intensity when underwater.
       Greatly improves shadows at shorelines.